src/requirements.txt
PyPDF2
pytesseract
pdf2image
requests
pillow
pandas
PyYAML
openai
genanki